version 0.3.2
Feat:
- Added context to the editpost page to only show the edit page if the userID matches the post authorID
- Added improved formatting to the login page
- Added context to the post page to only show the edit page button if the userID matches the post authorID
- Added context check to publish to ensure a user is logged in before viewing publish page

+++ front-end-app/src/components/editpost.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Form, Input, Button } from 'antd';
+import UserContext from '../contexts/user';
 
 const formItemLayout = {
   labelCol: { xs: { span: 24 }, sm: { span: 6 } },
@@ -36,12 +37,15 @@ class RegistrationForm extends React.Component {
     this.onFinish = this.onFinish.bind(this);
   }
 
+  static contextType = UserContext;
+
   componentDidMount() {
     const id = window.location.pathname.split('/')[2];
     fetch(`http://localhost:3030/TCS/listings/${id}`)
     .then(status)
     .then(json)
     .then(post => {
+      console.log(post)
       this.setState({post:post})
     })
     .catch(err => {
@@ -91,42 +95,46 @@ class RegistrationForm extends React.Component {
     }
     const post = this.state.post;
 
-
-    return (
-      <Form {...formItemLayout} name="Finalize" onFinish={this.onFinish} validateMessages={validateMessages} scrollToFirstError
-        initialValues={{
-          title: post.title,
-          breed: post.breed,
-          summary: post.summary,
-          imageURL: post.imageURL,
-        }}
-      >
-      
-        <Form.Item name="title" label="Listing Title" rules={[{ required: true }]}>     
-          <Input />      
-        </Form.Item>
+    if (this.context.user.ID === this.state.post.authorID) {
+      return (
+        <Form {...formItemLayout} name="Finalize" onFinish={this.onFinish} validateMessages={validateMessages} scrollToFirstError
+          initialValues={{
+            title: post.title,
+            breed: post.breed,
+            summary: post.summary,
+            imageURL: post.imageURL,
+          }}
+        >
+          <Form.Item name="title" label="Listing Title" rules={[{ required: true }]}>     
+            <Input />      
+          </Form.Item>
+          
+          <Form.Item name="breed" label="Dog Breed" rules={[{ required: true }]}>     
+            <Input />   
+          </Form.Item>
         
-        <Form.Item name="breed" label="Dog Breed" rules={[{ required: true }]}>     
-          <Input />   
-        </Form.Item>
-      
-        <Form.Item {...formSummaryLayout} name="summary" label="Summary" >      
-          <Input.TextArea /> 
-        </Form.Item>
-      
-        <Form.Item name="imageURL" label="Image URL?" >      
-          <Input />
-        </Form.Item>
-      
-        <Form.Item {...tailFormItemLayout}>     
-          <Button type="primary" htmlType="submit" >     
-            Finalize   
-          </Button> 
-          <Button type="primary" htmlType="submit" onClick={()=>this.deleteListing()} danger >     
-            Delete Listing  
-          </Button>  
-        </Form.Item >
-      </Form>     
+          <Form.Item {...formSummaryLayout} name="summary" label="Summary" >      
+            <Input.TextArea /> 
+          </Form.Item>
+        
+          <Form.Item name="imageURL" label="Image URL?" >      
+            <Input />
+          </Form.Item>
+        
+          <Form.Item {...tailFormItemLayout}>     
+            <Button type="primary" htmlType="submit" >     
+              Finalize   
+            </Button> 
+            <Button type="primary" htmlType="submit" onClick={()=>this.deleteListing()} danger >     
+              Delete Listing  
+            </Button>  
+          </Form.Item >
+        </Form>     
+      );
+    }
+    
+    return (
+      <p>Please login as the author of this listing to edit it</p>
     );
   };
 };
